Refactor, clean-up and address reviews of the first DNS over TCP PR
* Use RFC 1035 'bytes offset' to decode DNS over TCP payloads
* Correct Streams management
* Improve error management (for Debug and published Notes)
* Tests improvement
* Split files of ```package dns```

Minor changes:
* Change the name of dnsPrivateData to dnsConnectionData to reflect
the naming used in other applayers
* Split the ```Parse()``` method in multiple functions to comply more with the code convention
 used in other applayers implementation
* Remove a PCAP file from the previous and first DNS over TCP pull request
* Introduce a README.md file
